requirements.txt
streamlit
pandas
numpy
plotly
PyPortfolioOpt==1.5.6
cvxpy
scikit-learn
psycopg2-binary
sqlalchemy
python-dotenv
